# swagger.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntityDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"Get all users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntityDto"}}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/me":{"get":{"operationId":"UsersController_findCurrentUser","parameters":[],"responses":{"200":{"description":"Update user by ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntityDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/{id}":{"get":{"operationId":"UsersController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get user by ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntityDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Update user by ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntityDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Remove user by ID"}},"tags":["Users"],"security":[{"bearer":[]}]}},"/certificates/transfer":{"patch":{"operationId":"CertificatesController_transferCurrentUserCertificate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferCertificateDto"}}}},"responses":{"200":{"description":"Transfer my certificate","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CertificateEntityDto"}}}}},"tags":["Certificates"],"security":[{"bearer":[]}]}},"/certificates/my":{"get":{"operationId":"CertificatesController_findMyCerificates","parameters":[],"responses":{"200":{"description":"Find my certificates","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CertificateEntityDto"}}}}}},"tags":["Certificates"],"security":[{"bearer":[]}]}},"/certificates/available":{"get":{"operationId":"CertificatesController_findAvailable","parameters":[],"responses":{"200":{"description":"Available certificates","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CertificateEntityDto"}}}}}},"tags":["Certificates"],"security":[{"bearer":[]}]}},"/auth/local/signup":{"post":{"operationId":"AuthController_signupLocal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"User successfully registered and signed in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokensDto"}}}}},"tags":["Auth"]}},"/auth/local/signin":{"post":{"operationId":"AuthController_signinLocal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninLocalDto"}}}},"responses":{"200":{"description":"User successfully signed in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokensDto"}}}}},"tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"User successfully logged out"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/auth/refresh":{"post":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"200":{"description":"User tokens successfully refreshed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokensDto"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}}},"info":{"title":"Agreena API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:5000","description":"Agreena Local"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"}},"required":["firstName","lastName","username","email","phone","password"]},"UserEntityDto":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","firstName","lastName","username","email","phone","password","createdAt","updatedAt","deletedAt"]},"UpdateUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"}}},"TransferCertificateDto":{"type":"object","properties":{"ownerId":{"type":"string"},"certificateId":{"type":"string"}},"required":["ownerId","certificateId"]},"CertificateEntityDto":{"type":"object","properties":{"id":{"type":"string"},"country":{"type":"string"},"status":{"type":"string","enum":["Available","Owned","Transferred"]},"ownerId":{"type":"string"},"owner":{"$ref":"#/components/schemas/UserEntityDto"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","country","status","ownerId","owner","createdAt","updatedAt","deletedAt"]},"AuthTokensDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]},"SigninLocalDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}}}